+++ .travis.yml
@@ -15,10 +15,6 @@ notifications:
   email: false
 
 script:
-  - go build -buildmode=plugin -o plugins/logging_dispatcher_plugin.so plugins/logging_dispatcher/*.go
-  - go build -buildmode=plugin -o plugins/db_logging_plugin.so plugins/db_logging/*.go
-  - go build -buildmode=plugin -o plugins/graylog_logging_plugin.so plugins/graylog_logging/*.go
-  - go build -buildmode=plugin -o plugins/splunk_logging_plugin.so plugins/splunk_logging/*.go
   - go build -o bin/osctrl-tls cmd/tls/*.go
   - go build -o bin/osctrl-admin cmd/admin/*.go
   - go build -o bin/osctrl-cli cmd/cli/*.go
