+++ go.mod
@@ -3,15 +3,18 @@ module github.com/jmpsec/osctrl
 go 1.12
 
 require (
+	cloud.google.com/go v0.37.4 // indirect
 	github.com/beevik/etree v1.1.0 // indirect
 	github.com/crewjam/saml v0.0.0-20190508002657-ca21de9dd5b9
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gorilla/mux v1.7.2
 	github.com/gorilla/securecookie v1.1.1
 	github.com/gorilla/sessions v1.1.3
-	github.com/jinzhu/gorm v1.9.10
+	github.com/jinzhu/gorm v1.9.12
+	github.com/jmpsec/osctrl/pkg/backend v0.1.9
 	github.com/jmpsec/osctrl/pkg/carves v0.1.9
 	github.com/jmpsec/osctrl/pkg/environments v0.1.9
+	github.com/jmpsec/osctrl/pkg/logging v0.1.9
 	github.com/jmpsec/osctrl/pkg/metrics v0.1.9
 	github.com/jmpsec/osctrl/pkg/nodes v0.1.9
 	github.com/jmpsec/osctrl/pkg/queries v0.1.9
@@ -23,9 +26,11 @@ require (
 	github.com/olekukonko/tablewriter v0.0.1
 	github.com/russellhaering/goxmldsig v0.0.0-20180430223755-7acd5e4a6ef7 // indirect
 	github.com/segmentio/ksuid v1.0.2
-	github.com/spf13/viper v1.4.0
+	github.com/spf13/viper v1.6.2
 	github.com/urfave/cli v1.20.0
 )
+
+replace github.com/jmpsec/osctrl/pkg/backend => ./pkg/backend
 
 replace github.com/jmpsec/osctrl/pkg/carves => ./pkg/carves
 
@@ -45,10 +50,4 @@ replace github.com/jmpsec/osctrl/pkg/users => ./pkg/users
 
 replace github.com/jmpsec/osctrl/pkg/utils => ./pkg/utils
 
-replace github.com/jmpsec/osctrl/plugins/logging_dispatcher => ./plugins/logging_dispatcher
-
-replace github.com/jmpsec/osctrl/plugins/db_logging => ./plugins/db_logging
-
-replace github.com/jmpsec/osctrl/plugins/splunk_logging => ./plugins/splunk_logging
-
-replace github.com/jmpsec/osctrl/plugins/graylog_logging => ./plugins/graylog_logging
+replace github.com/jmpsec/osctrl/pkg/logging => ./pkg/logging
