+++ pkg/logging/go.mod
@@ -1,0 +1,8 @@
+module github.com/jmpsec/osctrl/pkg/logging
+
+go 1.12
+
+require (
+	github.com/jinzhu/gorm v1.9.12
+	github.com/spf13/viper v1.6.2
+)
