+++ pkg/metrics/go.mod
@@ -1,4 +1,4 @@
-module github.com/jmpsec/osctrl/metrics
+module github.com/jmpsec/osctrl/pkg/metrics
 
 go 1.12
